+++ src/ic-cdk-macros/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- `cargo build` should no longer give a confusing linkage error on Linux.
+
 ## [0.13.2] - 2024-04-08
 
 ### Changed